+++ tools/setup_pv.sh
@@ -25,40 +25,25 @@
 # - key-based SSH setup between the workstation and k8s master node
 #
 # Usage: on the workstation,
-# $ bash setup_pv.sh <setup|clean> <master> <username> <name> <path> <size> [storageClassName]
-#   setup|clean: setup or remove (including host files)
-#   master: IP address or hostname of k8s master node
-#   username: username on the server where the master was installed (this is
-#     the user who setup the cluster, and for which key-based SSH is setup)
-#   name: name of the PV, e.g. "pv-001"
-#   path: path of the host folder where 'name' should be created (if not existing)
-#   size: size in Gi to allocate to the PV
-#   storageClassName: storageClassName to assign
+#  $ bash setup_pv.sh <setup|clean|all> <path> <name> <size> <owner> [storageClassName]
+#    setup|clean|all: setup, remove (including host files), or both
+#    path: path of the host folder where 'name' should be created (if not existing)
+#    name: name of the PV, e.g. "pv-001"
+#    size: size in Gi to allocate to the PV
+#    owner: owner to set for the PV folder
+#    storageClassName: (optional) storageClassName to assign
 #
-
-function run_tmp() {
-  trap 'fail' ERR
-  if [[ "$master" != "$HOSTNAME"* ]]; then
-    ssh -x -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no \
-      $username@$master 'bash -s' < $tmp
-  else
-    bash $tmp
-  fi
-  rm $tmp
-}
 
 function setup() {
   trap 'fail' ERR
   # Per https://kubernetes.io/docs/tasks/configure-pod-container/configure-persistent-volume-storage/
-  tmp=/tmp/$(uuidgen)
-  cat <<EOF >$tmp
-if [[ ! \$(kubectl get pv $name) ]]; then
-  if [[ ! -e $path/$name ]]; then
-    sudo mkdir -p $path/$name
-    sudo chown \$USER:users $path/$name
-    chmod 777 $path/$name
-  fi
-  cat <<EOG >$name.yaml
+  if [[ -e $path/$name ]]; then sudo rm -rf $path/$name; fi
+  sudo mkdir -p $path/$name
+  sudo chown $owner $path/$name
+  sudo chmod 777 $path/$name
+  if [[ "$(kubectl get pv $name)" == "" ]]; then
+    local tmp=/tmp/$(uuidgen)
+    cat <<EOF >$tmp
 kind: PersistentVolume
 apiVersion: v1
 metadata:
@@ -73,39 +58,54 @@ spec:
   accessModes:
     - ReadWriteOnce
   hostPath:
-    path: "/$path/$name"
-EOG
-  kubectl create -f $name.yaml
-  kubectl get pv $name
-else
-  echo "PV $name already exists"
-  exit 1
-fi
+    path: "$path/$name"
 EOF
-  run_tmp
+
+    kubectl create -f $tmp
+    kubectl get pv $name
+    rm $tmp
+  else
+    log "WARN: PV $name already exists"
+  fi
 }
 
 function clean() {
   trap 'fail' ERR
-  tmp=/tmp/$(uuidgen)
-  cat <<EOF >$tmp
-kubectl delete pv $name
-sudo rm -rf /$path/$name
-EOF
-  run_tmp
+  cleanup_stuck_pvs
+  reset_pv_claim=""
+  clean_pv_data $name $path/$name
+  if [[ "$pv_claim" != "" ]]; then
+    if [[ "$pv_claim_refs" != "" ]]; then
+      log "WARN: PV $name is currently claimed and in use by pods. If needed, cleanup references first, e.g. via clean.sh"
+    else
+      namespace=$(echo $pv_claim_refs | jq -r ".[0].namespace")
+      if [[ "$(kubectl delete pvc -n $namespace $pv_claim)" ]]; then
+        log "PVC $pv_claim in namespace $namespace deleted"
+        kubectl delete pv $name
+        log "PV $name deleted"
+      else
+        log "WARN: PVC $pv_claim in namespace $namespace could not be deleted"
+      fi
+    fi
+  elif [[ "$(kubectl get pv $name)" != "" ]]; then
+    kubectl delete pv $name
+    log "PV $name deleted"
+  fi
+  if [[ -e $path/$name ]]; then
+    log "Deleting host folder $path/$name"
+    sudo rm -rf $path/$name
+  fi
 }
 
-if [[ $# -lt 6 ]]; then
+if [[ $# -lt 5 ]]; then
   cat <<'EOF'
- $ bash setup_pv.sh <setup|clean> <master> <username> <name> <path> <size> [storageClassName]
-   setup|clean: setup or remove (including host files)
-   master: IP address or hostname of k8s master node
-   username: username on the server where the master was installed (this is
-     the user who setup the cluster, and for which key-based SSH is setup)
-   name: name of the PV, e.g. "pv-001"
+ $ bash setup_pv.sh <setup|clean|all> <path> <name> <size> <owner> [storageClassName]
+   setup|clean|all: setup, remove (including host files), or both
    path: path of the host folder where 'name' should be created (if not existing)
+   name: name of the PV, e.g. "pv-001"
    size: size in Gi to allocate to the PV
-   storageClassName: storageClassName to assign
+   owner: owner to set for the PV folder
+   storageClassName: (optional) storageClassName to assign
 EOF
   echo "All parameters not provided"
   exit 1
@@ -118,17 +118,13 @@ cd $(dirname "$0")
 export AIO_ROOT="$(cd ../AIO; pwd -P)"
 source $AIO_ROOT/utils.sh
 action=$1
-master=$2
-username=$3
-name=$4
-path=$5
-size=$6
-storageClassName=$7
+path=$2
+name=$3
+size=$4
+owner=$5
+storageClassName=$8
 
-if [[ "$action" == "clean" ]]; then
-  clean
-else
-  setup
-fi
+if [[ "$action" == "clean" || "$action" = "all" ]]; then clean; fi
+if [[ "$action" == "setup" || "$action" = "all" ]]; then setup; fi
 
 cd $WORK_DIR
